GUI installer enhancements branch merging.
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@310 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/util/FileExecutor.java
@@ -141,60 +141,90 @@ public class FileExecutor
    * @param  output  Description of the Parameter
    * @return         Description of the Return Value
    */
-  public int executeCommand(String[] params, String[] output)
-  {
-    Process process = null;
-    MonitorInputStream outMonitor = null;
-    MonitorInputStream errMonitor = null;
-    Thread t1 = null;
-    Thread t2 = null;
-    int exitStatus = 0;
+	public int executeCommand(String[] params, String[] output)
+	{
+		StringBuffer retval = new StringBuffer();
+	    retval.append("executeCommand\n");
+	    if (params != null)
+	    {
+	    	for (int i = 0; i < params.length; i++)
+			{
+			    retval.append("\tparams: "+params[i]);
+			    retval.append("\n");
+			}
+	    }
+		Process process = null;
+		MonitorInputStream outMonitor = null;
+		MonitorInputStream errMonitor = null;
+		Thread t1 = null;
+		Thread t2 = null;
+		int exitStatus = -1;
 
-    try
-    {
-      // execute command
-      process = Runtime.getRuntime().exec(params);
+		try
+		{
+			// execute command
+			process = Runtime.getRuntime().exec(params);
 
-      InputStreamReader or = new InputStreamReader(process.getInputStream());
-      InputStreamReader er = new InputStreamReader(process.getErrorStream());
-      StringWriter outWriter = new StringWriter();
-      StringWriter errWriter = new StringWriter();
-      outMonitor = new MonitorInputStream(or, outWriter);
-      errMonitor = new MonitorInputStream(er, errWriter);
-      t1 = new Thread(outMonitor);
-      t2 = new Thread(errMonitor);
-      t1.setDaemon(true);
-      t2.setDaemon(true);
-      t1.start();
-      t2.start();
+            boolean console = false;//TODO: impl from xml <execute in_console=true ...>, but works already if this flag is true
+            if (console)
+            {
+    			Console c = new Console(process);
+    			// save command output
+    			output[0] = c.getOutputData();
+    			output[1] = c.getErrorData();
+            }
+            else
+            {
+    			StringWriter outWriter = new StringWriter();
+    			StringWriter errWriter = new StringWriter();
 
-      // wait for command to comlete
-      exitStatus = process.waitFor();
-      if (t1 != null)
-        t1.join();
+    			InputStreamReader or =
+    				new InputStreamReader(process.getInputStream());
+    			InputStreamReader er =
+    				new InputStreamReader(process.getErrorStream());
+    			outMonitor = new MonitorInputStream(or, outWriter);
+    			errMonitor = new MonitorInputStream(er, errWriter);
+    			t1 = new Thread(outMonitor);
+    			t2 = new Thread(errMonitor);
+    			t1.setDaemon(true);
+    			t2.setDaemon(true);
+    			t1.start();
+    			t2.start();
 
-      if (t2 != null)
-        t2.join();
+    			// wait for command to comlete
+    			exitStatus = process.waitFor();
+    			if (t1 != null)
+    			{
+    				t1.join();
+    			}
+    			if (t2 != null)
+    			{
+    				t2.join();
+    			}
 
-      // save command output
-      output[0] = outWriter.toString();
-      output[1] = errWriter.toString();
-      exitStatus = process.exitValue();
-    }
-    catch (InterruptedException e)
-    {
-      e.printStackTrace(System.err);
-      stopThread(t1, outMonitor);
-      stopThread(t2, errMonitor);
-      process.destroy();
-    }
-    catch (IOException e)
-    {
-      e.printStackTrace(System.err);
-    }
-    return exitStatus;
-  }
-
+    			// save command output
+    			output[0] = outWriter.toString();
+    			output[1] = errWriter.toString();
+            }
+			exitStatus = process.exitValue();
+		}
+		catch (InterruptedException e)
+		{
+			e.printStackTrace(System.err);
+			stopThread(t1, outMonitor);
+			stopThread(t2, errMonitor);
+			output[0] = "";
+			output[1] = e.getMessage() + "\n";
+			process.destroy();
+		}
+		catch (IOException e)
+		{
+			e.printStackTrace(System.err);
+			output[0] = "";
+			output[1] = e.getMessage() + "\n";
+		}
+		return exitStatus;
+	}
 
   /**
    *  Executes files specified at construction time.
@@ -213,24 +243,33 @@ public class FileExecutor
     Iterator efileIterator = files.iterator();
     while ((exitStatus == 0) && efileIterator.hasNext())
     {
+      boolean deleteAfterwards = true;
       ExecutableFile efile = (ExecutableFile) efileIterator.next();
       File file = new File(efile.path);
+      Debug.trace("handeling executable file "+efile);
 
       // fix executable permission for unix systems
       if (pathSep.equals(":") && (!osName.startsWith("mac") ||
         osName.endsWith("x")))
       {
+      	Debug.trace("making file executable (setting executable flag)");
         String[] params = {"/bin/chmod", permissions, file.toString()};
         exitStatus = executeCommand(params, output);
       }
       // loop through all operating systems
       Iterator osIterator = efile.osList.iterator();
+      if (!osIterator.hasNext())
+      {
+      	Debug.trace("no os to install the file on!");
+      }
       while (osIterator.hasNext())
       {
         Os os = (Os) osIterator.next();
 
+      	Debug.trace("checking if os param on file "+os+" equals current os");
         if (os.matchCurrentSystem())
         {
+	      	Debug.trace("match current os");
           // execute command in POSTINSTALL stage
           if ((exitStatus == 0) && (efile.executionStage == ExecutableFile.POSTINSTALL))
           {
@@ -263,6 +302,7 @@ public class FileExecutor
             // bring a dialog depending on return code and failure handling
             if (exitStatus != 0)
             {
+              deleteAfterwards = false;
               String message = output[0] + "\n" + output[1];
               if (message.length() == 1)
                 message = new String("Failed to execute " + file.toString() + ".");
@@ -294,12 +334,15 @@ public class FileExecutor
         }
         else
         {
-          //@todo
+	      	Debug.trace("-no match with current os!");
         }
       }
-      // POSTINSTALL executables will be deleted
-      if (efile.executionStage == ExecutableFile.POSTINSTALL)
-        file.delete();
+
+		// POSTINSTALL executables will be deleted
+		if (efile.executionStage == ExecutableFile.POSTINSTALL && deleteAfterwards)
+		{
+			if (file.canWrite()) file.delete();
+		}
 
     }
     return exitStatus;
